+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -6,7 +6,7 @@ package tech.skot
  * This code is generated from build.gradle.kts
  */
 public object Versions {
-  public const val skot: String = "1.4.0-ua"
+  public const val skot: String = "1.4.1-ua"
 
   public const val group: String = "com.github.useradgents.skot"